Refactor - Rename components to PascalCase
* Added dividers in sidebar
